+++ ports/rust/src/gui/widgets/mod.rs
@@ -1,1 +1,4 @@
 pub mod bcs_widget;
+pub mod bcs_window;
+pub mod bcs_button;
+pub mod bcs_label;